Reworks the Handle, still incomplete
Adds basic methods for sending and receiving events, now when a request
is sent, its id is also sent, if that one can't be used an event will
be sent to the handle.
Adds event NewId for when the id is already used, it informs the handler
of the newly selected one.
Renames RequestType in RequestKind (still not quite right).
The trait Map now has a private trait bound to Requestable which
implements the mapping-to-event transformation.